Added .dat files to .gitignore Modified plotting functions for better titling etc Added new MATLAB file to plot the generated cluster data. This is necessary because I can't get python to do 3D plotting on my Mac.

+++ ripser-investigation/Tools/histogramDistances.py
@@ -29,7 +29,7 @@ if __name__ == "__main__":
     for line in lines:
         words = line.split(',')
         words = words[0:-1]
-        temp = [int(word) for word in words]
+        temp = [float(word) for word in words]
         distances = distances + temp
 
     print "Read ",len(distances)," distances"
